Fixed tools and armor being unable to be repaired Fixed biome borders Fixed ichor being replaced by blood from another mod Fixed pigtificate trades not updating correctly Fixed enchanted books not having enchantments when read from a config Added nether fortresses to all biomes Added global biome configs Added global pigtificate trade configs Cleaned up some code

+++ src/main/java/logictechcorp/netherex/block/BlockEnokiMushroomStem.java
@@ -113,7 +113,7 @@ public class BlockEnokiMushroomStem extends BlockMod
         float maxY = state.getValue(UP) ? 1.0F : 0.8125F;
         float maxZ = state.getValue(SOUTH) ? 1.0F : 0.8125F;
 
-        return new AxisAlignedBB((double) minX, (double) minY, (double) minZ, (double) maxX, (double) maxY, (double) maxZ);
+        return new AxisAlignedBB(minX, minY, minZ, maxX, maxY, maxZ);
     }
 
     @Override
